Add bins for neighbors
This is necessary since line obstacles with vertices in the middle have visible
neighbors on both sides (e.g. above and below). To know which neighbor is on
which side is crucial to a network based routing algorithm. Otherwise A* or so
would route right through the obstacle, because there'll be an edge to and from
the vertices in the middle of the line obstacle.